% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_cov_canonical.R
\name{pois_cov_canonical}
\alias{pois_cov_canonical}
\title{Get List of Canonical Covariance Matrices}
\usage{
pois_cov_canonical(data)
}
\arguments{
\item{data}{\dQuote{pois.mash} data object, typically created by
calling \code{\link{pois_mash_set_data}}.}
}
\value{
A list of numerical vectors each of which forms a rank-1
  canonical covariance matrix.
}
\description{
Compute a list of canonical covariance matrices
  modeling condition-specific effects.
}
