% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_cov_ed.R
\name{pois_cov_ed}
\alias{pois_cov_ed}
\title{Estimate Prior Covariance Matrices using Extreme Deconvolution}
\usage{
pois_cov_ed(
  data,
  subset = NULL,
  Ulist,
  ulist,
  ulist.dd = NULL,
  ruv = FALSE,
  Fuv = NULL,
  verbose = FALSE,
  init = list(NULL),
  control = list(maxiter = 500, maxiter.q = 25, maxpsi2 = NULL, maxbias = 10, tol.stop
    = 1e-06, tol.q = 0.01, tol.rho = 1e-06)
)
}
\arguments{
\item{data}{\dQuote{pois.mash} data object, typically created by
calling \code{\link{pois_mash_set_data}}.}

\item{subset}{The indices of features to be used. Defaults to using
all features.}

\item{Ulist}{A list of H full-rank covariance matrices (e.g.,
initialized by \code{pois_cov_init}).}

\item{ulist}{A list of G numeric vectors each of which defines a
rank-1 covariance matrix.}

\item{ulist.dd}{Logical vector of length G denoting whether each
element in \code{ulist} is data-driven (\code{TRUE}) or canonical
(\code{FALSE}). Defaults to data-driven for all elements. For
canonical covariances, the spanned space is not updated.}

\item{ruv}{Logical scalar indicating whether to account for
unwanted variation. If \code{ruv = TRUE}, \code{Fuv} must be
provided.}

\item{Fuv}{J x D matrix of latent factors causing unwanted
variation, with features as rows and latent factors as columns.}

\item{verbose}{Logical scalar indicating whether to print ELBO at
each iteration.}

\item{init}{Optional list of initial values for model parameters
(e.g., returned by \code{pois_mash_ruv_prefit}).}

\item{control}{A list of control parameters with the following
elements: \item{maxiter}{Maximum number of ED iterations. Default
is 500.}  \item{maxiter.q}{Maximum number of inner loop iterations
to update variational parameters at each ED iteration. Default is
25.}  \item{maxpsi2}{Maximum for the gene-specific dispersion
parameter psi2.}  \item{maxbias}{Maximum for the gene-specific
range of bias caused by unwanted variation. Default is 10.}
\item{tol.stop}{Tolerance for assessing convergence of ED, as
measured by relative change in ELBO. Default is 1e-6.}
\item{tol.q}{Relative tolerance for assessing convergence of
variational parameters at each ED iteration. Default is 1e-2.}
\item{tol.rho}{Tolerance for assessing convergence of effects
corresponding to unwanted variation. Default is 1e-6.}}
}
\value{
A list including the following elements:

\item{Ulist}{A list of H full-rank covariance matrices.}

\item{ulist}{A list of G numeric vectors each of which forming a
rank-1 covariance matrix.}

\item{pi}{(H+G) by 1 numeric vector of mixture proportions for
Ulist and ulist.}
}
\description{
Perform extreme deconvolution (ED) to estimate
  data-driven prior covariance matrices.
}
