% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_cov_init.R
\name{pois_cov_init}
\alias{pois_cov_init}
\title{Initialize data-driven prior covariance matrices based on principal component analysis}
\usage{
pois_cov_init(
  data,
  ruv = FALSE,
  Fuv = NULL,
  rho = NULL,
  prop = 1,
  seed = 1,
  npc = 5,
  cutoff = 3
)
}
\arguments{
\item{data}{A pois.mash data object containing the following
components: X, a J x R matrix of count data collapsed over
conditions, with features as rows and conditions as columns; s, an
R x 1 numeric vector adjusting for sequencing depth of each of R
conditions; subgroup, an R x 1 factor vector with M levels denoting
the subgroup status of each of R conditions.}

\item{ruv}{A logical scalar indicating whether to account for unwanted variation. Default is \code{FALSE}. 
If \code{TRUE}, Fuv and rho must be provided.}

\item{Fuv}{A J by D matrix of latent factors causing unwanted variation, with features as rows and latent factors as columns}

\item{rho}{A D by R matrix of effects corresponding to unwanted variation, such that bias = \code{Fuv \%*\% rho}}

\item{prop}{The proportion by which to take a random subset of genes for prior covariance estimation (useful in case of many genes). Default is 1.}

\item{seed}{Specified for reproducibility if prop is strictly smaller than 1. Default is 1.}

\item{npc}{The number of principal components to use. Default is 5.}

\item{cutoff}{The threshold for the maximum of absolute values of Z-scores taken across conditions to include as "strong" features 
used for prior covariance estimation. Default is 3.}
}
\value{
A list with initial estimates of prior covariances, and indices of features (from j=1, ..., J) to include in the subsequent ED step
}
\description{
Initialize data-driven prior covariance matrices based on principal component analysis
}
