% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mash_ruv_prefit.R
\name{pois_mash_ruv_prefit}
\alias{pois_mash_ruv_prefit}
\title{Prefit the poisson mash model to get an initial estimate of model parameters if accounting for unwanted variation}
\usage{
pois_mash_ruv_prefit(
  data,
  Fuv,
  verbose = FALSE,
  init = list(NULL),
  control = list(maxiter = 100, maxiter.q = 25, tol.q = 0.01, tol.rho = 0.0001,
    tol.stop = 1e-06)
)
}
\arguments{
\item{data}{A pois.mash data object containing the following
components: X, J x R matrix of count data collapsed over
conditions, with features as rows and conditions as columns; s, R x
1 numeric vector adjusting for sequencing depth of each of R
conditions; subgroup, R x 1 factor vector with M levels denoting
the subgroup status of each of R conditions.}

\item{Fuv}{J by D matrix of latent factors causing unwanted variation, with features as rows and latent factors as columns}

\item{verbose}{A logical scalar indicating whether to print ELBO at each iteration. Default is \code{FALSE}.}

\item{init}{A list of initial values for model parameters which could be empty}

\item{control}{A list of control parameters with the following
elements: maxiter, Maximum number of iterations. Default is 100;
maxiter.q, Maximum number of inner loop iterations to update
variational parameters at each iteration. Default is 25; tol.stop,
Tolerance for assessing convergence of prefit, as measured by
relative change in ELBO. Default is 1e-6; tol.q, Relative tolerance
for assessing convergence of variational parameters at each
iteration. Default is 1e-2; tol.rho, Tolerance for assessing
convergence of effects corresponding to unwanted variation. Default
is 1e-4.}
}
\value{
A list with initial estimates of model parameters
}
\description{
There is no need to call this function if unwanted variation is not accounted for.
}
