% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data.R
\name{pois_mash_set_data}
\alias{pois_mash_set_data}
\title{Create Data Oject for Poisson Mash Analysis}
\usage{
pois_mash_set_data(
  Y,
  condition,
  si = colSums(Y),
  subgroup = gl(1, nlevels(condition))
)
}
\arguments{
\item{Y}{J x N matrix of counts (non-negative numbers) with
features (e.g., genes) as rows and observations (e.g., cells) as
columns. Y should have at least 2 rows and 2 columns.}

\item{condition}{Factor of length N with R > 1 levels giving the
condition for each observation.}

\item{si}{Numeric vector of length N containing the size factors for
the N observations.}

\item{subgroup}{Named vector of factors with M levels denoting the
subgroup status of each of R conditions.  Default to no subgroup
(M=1). If not set to \code{NULL}, the names of subgroup must match
the levels of condition.}
}
\value{
A pois.mash data object for poisson mash analysis. It is a list
  with the following components:

\item{X}{J x R matrix of count data collapsed over conditions, with
  features as rows and conditions as columns.}

\item{s}{Vector of \dQuote{sequencing depths} for the R conditions.}

\item{subgroup}{R x 1 factor vector with M levels denoting the
  subgroup status of each of R conditions.}
}
\description{
Prepare count data for an analysis with Poisson mash.
}
\examples{
# Add examples here.

}
\seealso{

}
