% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{pois_mash_sim_data}
\alias{pois_mash_sim_data}
\title{Simulate Data Based on a Real Single-cell Dataset}
\usage{
pois_mash_sim_data(J = 1000, R = 5, effects, seed = 1)
}
\arguments{
\item{J}{Number of features to simulate (should be 2 or greater).}

\item{R}{Number of conditions to simulate (should be 2 or greater).}

\item{effects}{List consisting of the following components:
\dQuote{umat}, an R x K matrix with each of the K columns
specifying a pattern of non-null effects; \dQuote{prob}, a
non-negative vector of length K specifying the relative frequency
of each non-null effect among all J features.}

\item{seed}{Random seed (for reproducing the output).}
}
\value{
A list containing the following components:

\item{Y}{J x N sparse matrix of counts with features (e.g., genes)
  as rows and observations (e.g., cells) as columns.}

\item{condition}{Factor of length N with R levels giving the
  condition for each observation.}
}
\description{
Simulate Data Based on a Real Single-cell Dataset
}
\examples{
umat <- cbind(c(1,1,0,0,0),c(0,0,0,1,0.5))
pois_mash_sim_data(1000,5,effects = list(umat = umat,prob = rep(0.05,2)))

}
