% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{update_q_eta_only}
\alias{update_q_eta_only}
\title{Utility functions to update model and variational parameters and speed up matrix inversions}
\usage{
update_q_eta_only(
  x,
  s,
  mu,
  bias,
  c2,
  psi2,
  init = list(NULL),
  control = list(maxiter = 25, tol = 0.01, lwr = -10, upr = 10)
)
}
\description{
These are internal functions which users might not want to call directly.
}
