% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_cov_canonical.R
\name{pois_cov_canonical}
\alias{pois_cov_canonical}
\title{Generate Condition-specific Canonical Covariance Matrices}
\usage{
pois_cov_canonical(dat)
}
\arguments{
\item{data}{A Poisson mash data object, typically created using
\code{\link{pois_mash_set_data}}.}
}
\value{
A list ii which each list element is a vector \code{u}
defining a rank-1 covariance matrix \code{tcrossprod(u)}
}
\description{
Generate Condition-specific Canonical Covariance Matrices
}
