% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_cov_canonical.R
\name{pois_cov_canonical}
\alias{pois_cov_canonical}
\title{Compute a list of canonical covariance matrices modeling condition-specific effects}
\usage{
pois_cov_canonical(data)
}
\arguments{
\item{data}{A pois.mash data object containing the following components:
\item{X}{J x R matrix of count data collapsed over conditions, with features as rows and conditions as columns.} 
\item{s}{R x 1 numeric vector adjusting for sequencing depth of each of R conditions.}
\item{subgroup}{R x 1 factor vector with M levels denoting the subgroup status of each of R conditions.}}
}
\value{
a list of numerical vectors each of which forming a rank-1 canonical covariance matrix
}
\description{
Compute a list of canonical covariance matrices modeling condition-specific effects
}
