% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_cov_ed.R
\name{pois_cov_ed}
\alias{pois_cov_ed}
\title{Perform extreme deconvolution (ED) to estimate data-driven prior covariance matrices}
\usage{
pois_cov_ed(
  data,
  subset = NULL,
  Ulist,
  ulist,
  ulist.dd = NULL,
  ruv = FALSE,
  Fuv = NULL,
  verbose = FALSE,
  init = list(NULL),
  control = list(maxiter = 500, maxiter.q = 25, tol.q = 0.01, tol.rho = 1e-06, tol.stop
    = 1e-06)
)
}
\arguments{
\item{data}{A pois.mash data object containing the following
components: X, a J x R matrix of count data collapsed over
conditions, with features as rows and conditions as columns; s, an
R x 1 numeric vector adjusting for sequencing depth of each of R
conditions and subgroup, an R x 1 factor vector with M levels
denoting the subgroup status of each of R conditions.}

\item{subset}{The indices of features to be used. Defaults to all of them.}

\item{Ulist}{A list of H full-rank covariance matrices (e.g., initialized by \code{pois_cov_init})}

\item{ulist}{A list of G numeric vectors each of which forming a rank-1 covariance matrix}

\item{ulist.dd}{A G by 1 logical vector denoting whether each element in ulist is data-driven (\code{TRUE}) or canonical (\code{FALSE}). 
Defaults to data-driven for all elements. For canonical covariances, the spanned space is not updated.}

\item{ruv}{A logical scalar indicating whether to account for unwanted variation. Default is \code{FALSE}. If \code{TRUE}, Fuv must be provided.}

\item{Fuv}{A J by D matrix of latent factors causing unwanted variation, with features as rows and latent factors as columns}

\item{verbose}{A logical scalar indicating whether to print ELBO at each iteration. Default is \code{FALSE}.}

\item{init}{A list of initial values for model parameters (e.g., returned by \code{pois_mash_ruv_prefit}). Could be empty.}

\item{control}{A list of control parameters with the following
elements: maxiter, maximum number of ED iterations. Default is 500;
maxiter.q, maximum number of inner loop iterations to update
variational parameters at each ED iteration. Default is 25;
tol.stop, Tolerance for assessing convergence of ED, as measured by
relative change in ELBO. Default is 1e-6; tol.q, relative tolerance
for assessing convergence of variational parameters at each ED
iteration. Default is 1e-2; tol.rho, Tolerance for assessing
convergence of effects corresponding to unwanted variation. Default
is 1e-6.}
}
\value{
A list with the following elements: 
\item{Ulist}{A list of H full-rank covariance matrices.}
\item{ulist}{A list of G numeric vectors each of which forming a rank-1 covariance matrix.}
\item{pi}{(H+G) by 1 numeric vector of mixture proportions for Ulist and ulist.}
}
\description{
Perform extreme deconvolution (ED) to estimate data-driven prior covariance matrices
}
