% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_cov_ed.R
\name{pois_cov_ed}
\alias{pois_cov_ed}
\alias{pois_cov_ed_control_default}
\title{Estimate Prior Covariance Matrices using Extreme Deconvolution}
\usage{
pois_cov_ed(
  data,
  subset,
  Ulist,
  ulist,
  ulist.dd,
  ruv = FALSE,
  Fuv,
  update.mu = TRUE,
  verbose = FALSE,
  init = list(),
  control = list()
)

pois_cov_ed_control_default()
}
\arguments{
\item{data}{\dQuote{pois.mash} data object, typically created by
calling \code{\link{pois_mash_set_data}}.}

\item{subset}{The indices of features to be used. Defaults to using
all features.}

\item{Ulist}{A list of H full-rank covariance matrices (e.g.,
initialized by \code{pois_cov_init}).}

\item{ulist}{A list of G numeric vectors each of which defines a
rank-1 covariance matrix.}

\item{ulist.dd}{Logical vector of length G denoting whether each
element in \code{ulist} is data-driven (\code{TRUE}) or canonical
(\code{FALSE}). Defaults to data-driven for all elements. For
canonical covariances, the spanned space is not updated.}

\item{ruv}{Logical scalar indicating whether to account for
unwanted variation. If \code{ruv = TRUE}, \code{Fuv} must be
provided.}

\item{Fuv}{J x D matrix of latent factors causing unwanted
variation, with features as rows and latent factors as columns.}

\item{update.mu}{A logical scalar indicating whether to update 
gene-specific means mu. If \code{update.mu = FALSE}, initial mu
must be provided in \code{init}.}

\item{verbose}{Logical scalar indicating whether to print ELBO at
each iteration.}

\item{init}{Optional list of initial values for model parameters
(e.g., returned by \code{pois_mash_prefit}).}

\item{control}{A list of control parameters with the following
elements: \dQuote{maxiter}, maximum number of ED iterations;
\dQuote{maxiter.q}, maximum number of inner loop iterations to
update variational parameters at each ED iteration;
\dQuote{maxpsi2}, maximum for the gene-specific dispersion
parameter \code{psi2}; \dQuote{maxbias}, maximum for the
gene-specific range of bias caused by unwanted variation;
\dQuote{tol.stop}, tolerance for assessing convergence of ED, as
measured by absolute change in ELBO; \dQuote{tol.q}, relative
tolerance for assessing convergence of variational parameters at
each ED iteration; and \dQuote{tol.rho}{tolerance for assessing
convergence of effects corresponding to unwanted variation.} Any
named components will override the default optimization algorithm
settings (as they are defined by
\code{pois_cov_ed_control_default}).}
}
\value{
A list including the following elements:

\item{Ulist}{List of H full-rank covariance matrices.}

\item{ulist}{List of G numeric vectors each of which forms a
  rank-1 covariance matrix.}

\item{pi}{Numeric vector of length H + G containing the mixture
  proportions for Ulist and ulist.}
}
\description{
Perform extreme deconvolution (ED) to estimate
  data-driven prior covariance matrices.
}
