% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_cov_init.R
\name{pois_cov_init}
\alias{pois_cov_init}
\title{Initialize Data-driven Prior Covariance Matrices using PCA}
\usage{
pois_cov_init(
  data,
  ruv = FALSE,
  Fuv = NULL,
  rho = NULL,
  prop = 1,
  seed = 1,
  npc = 5,
  cutoff = 3
)
}
\arguments{
\item{data}{A \dQuote{pois.mash} data object, typically created by
calling \code{\link{pois_mash_set_data}}.}

\item{ruv}{A logical scalar indicating whether to account for
unwanted variation. Default is \code{FALSE}. If \code{ruv = TRUE},
\code{Fuv} and \code{rho} must be provided.}

\item{Fuv}{J x D matrix of latent factors causing unwanted
variation, with features as rows and latent factors as columns.}

\item{prop}{The proportion by which to take a random subset of
genes for prior covariance estimation (useful in case of many
genes).}

\item{seed}{Useful for reproducibility when \code{prop} is less
than 1.}

\item{npc}{The number of principal components to use.}

\item{cutoff}{The threshold for the maximum of absolute values of
 Z-scores taken across conditions to include as "strong" features
used for prior covariance estimation.}
}
\value{
A list with initial estimates of prior covariances, and
  indices of the features (j = 1,...,J) to include in the subsequent
  ED step.
}
\description{
Initialize data-driven prior covariance matrices based
  on principal component analysis.
}
