% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mash_posterior.R
\name{pois_mash_posterior}
\alias{pois_mash_posterior}
\title{Compute Posterior Summaries From Poisson Mash Fit}
\usage{
pois_mash_posterior(
  data,
  s,
  mu,
  psi2,
  bias = matrix(0, nrow(data), ncol(data)),
  wlist,
  Ulist,
  ulist,
  ulist.epsilon2 = rep(1e-08, length(ulist)),
  zeta,
  thresh = 1/(500 * ncol(zeta)),
  C = diag(ncol(data)) - 1/ncol(data),
  res.colnames = paste0(colnames(data), "-mean"),
  posterior_samples = 0,
  median_deviations = FALSE,
  seed = 1
)
}
\arguments{
\item{data}{J x R matrix of counts collapsed over conditions, with
features as rows and conditions as columns.}

\item{s}{Numeric vector of length R adjusting for sequencing depth of
each of R conditions.}

\item{mu}{J x R matrix of gene-specific means (R conditions are
assumed to belong to M subgroups, so each row should have at most M
distinct values).}

\item{psi2}{Vector of length J giving the gene-specific dispersion
parameters.}

\item{bias}{J x R matrix of bias caused by unwanted
variation. Default is a matrix of all zeros.}

\item{wlist}{Numeric vector of length L specifying the scaling
factors for the prior covariance matrices.}

\item{Ulist}{List of H full-rank covariance matrices}

\item{ulist}{List of G numeric vectors each of which forms a
rank-1 covariance matrix.}

\item{ulist.epsilon2}{Numeric vector of length G added to the
diagonals of each rank-1 prior covariance matrix.}

\item{zeta}{J x K matrix of posterior weights, where K = L*(H + G) is
the number of prior mixture components.}

\item{thresh}{Posterior weights thresholds. Below this threshold, the
mixture components are ignored in the posterior calculations.}

\item{C}{Q x R matrix of contrasts for effects.}

\item{res.colnames}{Character vector of length Q giving the names
of the contrasts.}

\item{posterior_samples}{The number of samples to be drawn from the
posterior distribution of each effect.}

\item{median_deviations}{Logical scalar indicating whether to calculate
posterior summary of deviation of condition-specific effects 
relative to the median over all conditions.}

\item{seed}{a random number seed to use when sampling from the posteriors.
It is used when \code{posterior_samples > 0} 
or \code{median_deviations = TRUE}.}
}
\value{
The return value is a list with the following components:

\item{PosteriorMean}{J x Q matrix of posterior means.}

\item{PosteriorSD}{J x Q matrix of posterior standard deviations.}

\item{ZeroProb}{J x Q matrix in which each entry is the posterior
  probability of the true effect being zero.}

\item{NegativeProb}{J x Q matrix in which each entry is the
  posterior probability of the true effect being negative.}

\item{lfsr}{J x Q matrix of local false sign rate estimates.}

\item{PosteriorSamples}{J x R x posterior_samples array of samples of effects, 
if \code{posterior_samples > 0}.}

\item{beta_median_dev_post}{a list containing posterior summary of 
deviation of effects relative to the median, 
if \code{median_deviations = TRUE}.}
}
\description{
Compute posterior summaries of the given contrasts of
  the matrix of effects based on the poisson mash fit. Mixture
  components with very small posterior weights are ignored in the
  posterior calculations.

  Note this is an internal function which users might not want to
  call directly.
}
\keyword{internal}
