% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mash_posterior.R
\name{pois_mash_posterior}
\alias{pois_mash_posterior}
\title{Compute posterior summaries of the given contrasts of the matrix of effects based on the poisson mash fit}
\usage{
pois_mash_posterior(
  data,
  s,
  mu,
  psi2,
  bias = NULL,
  wlist,
  Ulist,
  ulist,
  ulist.epsilon2 = NULL,
  zeta,
  thresh = NULL,
  C = NULL,
  res.colnames = NULL
)
}
\arguments{
\item{data}{J by R matrix of counts collapsed over conditions, with features as rows and conditions as columns}

\item{s}{R by 1 numeric vector s adjusting for sequencing depth of each of R conditions}

\item{mu}{J by R matrix of gene-specific means (R conditions are assumed to belong to M subgroups, so each row should have at most M distinct values)}

\item{psi2}{J by 1 vector of gene-specific dispersion parameters}

\item{bias}{J by R matrix of bias caused by unwanted variation. Default to matrix of all zeros.}

\item{wlist}{L by 1 numeric vector of scaling factors for the prior covariance matrices}

\item{Ulist}{A list of H full-rank covariance matrices}

\item{ulist}{A list of G numeric vectors each of which forming a rank-1 covariance matrix}

\item{ulist.epsilon2}{G by 1 numeric vector that adds a small positive epsilon2 to the diagonals of each rank-1 prior covariance matrix}

\item{zeta}{J by K matrix of posterior weights, where K=L*(H+G) is the number of mixture components in the prior}

\item{thresh}{the threshold for posterior weights below which the corresponding mixture components are ignored in posterior summary calculation}

\item{C}{Q by R matrix of contrasts for effects}

\item{res.colnames}{Q by 1 character vector giving the names of the contrasts}
}
\value{
A list with the following components:
\item{PosteriorMean}{J x Q matrix of posterior means.}
\item{PosteriorSD}{J x Q matrix of posterior standard deviations.}
\item{ZeroProb}{J x Q matrix of posterior probability of being zero.}
\item{NegativeProb}{J x Q matrix of posterior probability of being negative.}
\item{lfsr}{J x Q matrix of local false sign rate estimates.}
}
\description{
This is an internal function which users might not want to call directly. 
Mixtures components with very small posterior weights are ignored in the calculation of effect posterior summaries.
}
\keyword{internal}
