% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mash_posterior.R
\name{pois_mash_posterior}
\alias{pois_mash_posterior}
\title{Compute posterior summaries of the given contrasts of the matrix of effects based on the poisson mash fit}
\usage{
pois_mash_posterior(
  data,
  s,
  mu,
  psi2,
  ruv = FALSE,
  Fuv = NULL,
  rho = NULL,
  wlist,
  Ulist,
  ulist,
  ulist.epsilon2 = NULL,
  zeta,
  C = NULL,
  res.colnames = NULL
)
}
\arguments{
\item{data}{J by R matrix of counts collapsed over conditions, with features as rows and conditions as columns}

\item{s}{R by 1 numeric vector s adjusting for sequencing depth of each of R conditions}

\item{mu}{J by R matrix of gene-specific means (R conditions are assumed to belong to M subgroups, so each row should have at most M distinct values)}

\item{psi2}{J by 1 vector of gene-specific dispersion parameters}

\item{ruv}{Logical scalar indicating whether to account for unwanted variation. Default is \code{FALSE}. 
If \code{TRUE}, Fuv and rho must be provided.}

\item{Fuv}{J by D matrix of latent factors causing unwanted variation, with features as rows and latent factors as columns}

\item{rho}{D by R matrix of effects corresponding to unwanted variation, such that bias = Fuv %*% rho}

\item{wlist}{L by 1 numeric vector of scaling factors for the prior covariance matrices}

\item{Ulist}{A list of H full-rank covariance matrices}

\item{ulist}{A list of G numeric vectors each of which forming a rank-1 covariance matrix}

\item{ulist.epsilon2}{G by 1 numeric vector that adds a small positive epsilon2 to the diagonals of each rank-1 prior covariance matrix}

\item{zeta}{J by K matrix of posterior weights, where K=L*(H+G) is the number of mixture components in the prior}

\item{C}{Q by R matrix of contrasts for effects}

\item{res.colnames}{Q by 1 character vector giving the names of the contrasts}
}
\value{
A list with the following components:
\item{PosteriorMean}{J x Q matrix of posterior means.}
\item{PosteriorSD}{J x Q matrix of posterior standard deviations.}
\item{ZeroProb}{J x Q matrix of posterior probability of being zero.}
\item{NegativeProb}{J x Q matrix of posterior probability of being negative.}
\item{lfsr}{J x Q matrix of local false sign rate estimates.}
}
\description{
This is an internal function which users might not want to call directly.
}
