% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mash_ruv_prefit.R
\name{pois_mash_ruv_prefit}
\alias{pois_mash_ruv_prefit}
\alias{pois_mash_ruv_prefit_control_default}
\title{Prefit Poisson MASH Model Accounting for Unwanted Variation}
\usage{
pois_mash_ruv_prefit(
  data,
  Fuv,
  update.mu = TRUE,
  verbose = FALSE,
  init = list(),
  control = list()
)

pois_mash_ruv_prefit_control_default()
}
\arguments{
\item{data}{\dQuote{pois.mash} data object, typically created by
calling \code{\link{pois_mash_set_data}}.}

\item{Fuv}{J x D matrix of latent factors causing unwanted
variation, with features as rows and latent factors as columns.}

\item{update.mu}{A logical scalar indicating whether to update 
gene-specific means mu. If \code{update.mu = FALSE}, initial mu
must be provided in \code{init}.}

\item{verbose}{Logical scalar indicating whether to print ELBO at
each iteration.}

\item{init}{List of initial values for model parameters which
could be empty.}

\item{control}{List of control parameters with one or more of the
following elements: \dQuote{maxiter}, maximum number of iterations;
\dQuote{maxiter.q}, maximum number of inner-loop iterations to
update variational parameters at each iteration; \dQuote{tol.stop},
tolerance for assessing convergence of prefit, as measured by
absolute change in ELBO; \dQuote{tol.q}, relative tolerance for
assessing convergence of variational parameters at each iteration;
\dQuote{tol.rho}, tolerance for assessing convergence of effects
corresponding to unwanted variation. Any named components will
override the default optimization algorithm settings (as they are
defined by \code{pois_mash_ruv_prefit_control_default}).}
}
\value{
A list containing initial estimates of model parameters.
}
\description{
Prefit the poisson mash model to get an initial
  estimate of model parameters if accounting for unwanted
  variation. Don't call this function if unwanted variation is not
  accounted for.
}
