% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_mash_ruv_prefit.R
\name{pois_mash_ruv_prefit}
\alias{pois_mash_ruv_prefit}
\title{Prefit Poisson MASH Model Accounting for Unwanted Variation}
\usage{
pois_mash_ruv_prefit(
  data,
  Fuv,
  verbose = FALSE,
  init = list(NULL),
  control = list(maxiter = 100, maxiter.q = 25, tol.q = 0.01, tol.rho = 0.0001,
    tol.stop = 1e-06)
)
}
\arguments{
\item{data}{A \dQuote{pois.mash} data object, typically created by
calling \code{\link{pois_mash_set_data}}.}

\item{Fuv}{J x D matrix of latent factors causing unwanted
variation, with features as rows and latent factors as columns.}

\item{verbose}{A logical scalar indicating whether to print ELBO at
each iteration.}

\item{init}{A list of initial values for model parameters which
could be empty.}

\item{control}{A list of control parameters with the following
elements: \dQuote{maxiter}, maximum number of iterations;
\dQuote{maxiter.q}, maximum number of inner-loop iterations to
update variational parameters at each iteration; \dQuote{tol.stop},
tolerance for assessing convergence of prefit, as measured by
relative change in ELBO; \dQuote{tol.q}, relative tolerance for
assessing convergence of variational parameters at each iteration;
\dQuote{tol.rho}, tolerance for assessing convergence of effects
corresponding to unwanted variation.}
}
\value{
A list containing initial estimates of model parameters.
}
\description{
Prefit the poisson mash model to get an initial
  estimate of model parameters if accounting for unwanted
  variation. Don't call this function if unwanted variation is not
  accounted for.
}
