% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data.R
\name{pois_mash_set_data}
\alias{pois_mash_set_data}
\title{Create a data object for poisson mash analysis}
\usage{
pois_mash_set_data(Y, condition, si, subgroup = NULL)
}
\arguments{
\item{Y}{A J by N matrix of counts with features (e.g., genes) as
rows and observations (e.g., cells) as columns}

\item{condition}{An N by 1 vector of factors with R levels denoting
the condition status of each observation}

\item{si}{An N by 1 vector of size factors for each observation}

\item{subgroup}{An R by 1 named vector of factors with M levels
denoting the subgroup status of each of R conditions.  Default to
no subgroup (M=1). If not set to \code{NULL}, the names of subgroup
must match the levels of condition.}
}
\value{
A pois.mash data object for poisson mash analysis,
including the following components:

\item{X}{J x R matrix of count data collapsed over conditions, with
  features as rows and conditions as columns.}

\item{s}{R x 1 numeric vector adjusting for sequencing depth of
  each of R conditions.}

\item{subgroup}{R x 1 factor vector with M levels denoting the
  subgroup status of each of R conditions.}
}
\description{
Create a data object for poisson mash analysis
}
