% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data.R
\name{pois_mash_set_data}
\alias{pois_mash_set_data}
\title{Create Data Object for Poisson MASH Analysis}
\usage{
pois_mash_set_data(Y, condition, si, subgroup)
}
\arguments{
\item{Y}{J x N matrix of counts with features (e.g., genes) as
rows and observations (e.g., cells) as columns.}

\item{condition}{Factor with N entries and R levels in which
condition[i] specifies the condition status of the ith observation.}

\item{si}{Vector of size factors for each observation.}

\item{subgroup}{Factor with N entries and M levels giving the
subgroup status of each of R conditions. Default is no subgroups
(that is, M = 1 level). If not set to \code{NULL}, the names of
subgroup must match the levels of \code{condition}.}
}
\value{
A \dQuote{pois.mash} data object for poisson mash analysis,
  including the following components:

\item{X}{J x R matrix of count data collapsed over conditions, with
  features as rows and conditions as columns.}

\item{s}{Vector of length R adjusting for sequencing depth of
  each of R conditions.}

\item{subgroup}{Vector with M levels specifying the subgroup status
  of each of R conditions.}
}
\description{
Create Data Object for Poisson MASH Analysis
}
